move connection manager to corehq.db
