+++ corehq/apps/userreports/pillow.py
@@ -2,8 +2,9 @@ from alembic.autogenerate.api import compare_metadata
 from casexml.apps.case.models import CommCareCase
 from corehq.apps.userreports.exceptions import TableRebuildError
 from corehq.apps.userreports.models import DataSourceConfiguration, CustomDataSourceConfiguration
-from corehq.apps.userreports.sql import create_engine, IndicatorSqlAdapter, metadata
+from corehq.apps.userreports.sql import IndicatorSqlAdapter, metadata
 from corehq.apps.userreports.tasks import rebuild_indicators
+from corehq.db import connection_manager
 from dimagi.utils.logging import notify_error
 from fluff.signals import get_migration_context, get_tables_to_rebuild
 from pillowtop.couchdb import CachedCouchDB
@@ -24,7 +25,7 @@ class ConfigurableIndicatorPillow(PythonPillow):
         # todo: switch to connection_manager
         engine = getattr(cls, '_engine', None)
         if not engine:
-            cls._engine = create_engine()
+            cls._engine = connection_manager.get_engine()
         return cls._engine
 
     def get_all_configs(self):
