+++ corehq/apps/userreports/sql/__init__.py
@@ -1,4 +1,3 @@
 from .adapter import get_indicator_table, IndicatorSqlAdapter, metadata
 from .columns import get_expanded_column_config, SqlColumnConfig
-from .connection import create_engine
 from .util import get_column_name, get_table_name, truncate_value
